+++ src/main/java/at/hannibal2/skyhanni/config/features/DamageIndicator.java
@@ -1,9 +1,9 @@
 package at.hannibal2.skyhanni.config.features;
 
-import at.hannibal2.skyhanni.config.gui.core.config.annotations.ConfigEditorBoolean;
-import at.hannibal2.skyhanni.config.gui.core.config.annotations.ConfigEditorDraggableList;
-import at.hannibal2.skyhanni.config.gui.core.config.annotations.ConfigEditorDropdown;
-import at.hannibal2.skyhanni.config.gui.core.config.annotations.ConfigOption;
+import at.hannibal2.skyhanni.config.core.config.annotations.ConfigEditorBoolean;
+import at.hannibal2.skyhanni.config.core.config.annotations.ConfigEditorDraggableList;
+import at.hannibal2.skyhanni.config.core.config.annotations.ConfigEditorDropdown;
+import at.hannibal2.skyhanni.config.core.config.annotations.ConfigOption;
 import com.google.gson.annotations.Expose;
 
 import java.util.ArrayList;