+++ src/main/java/at/hannibal2/skyhanni/config/gui/core/config/Config.java
@@ -1,5 +1,0 @@
-package at.hannibal2.skyhanni.config.gui.core.config;
-//public class Config {
-//
-//    public void executeRunnable(String runnableId) {}
-//}